requirements.txt
celery==5.2.7
Django==4.2
django-celery-beat==2.5.0
django-oauth-toolkit==2.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python3-logstash==0.4.80
